Core: Updated to 11.0.2
Port From (https://github.com/TrinityCore/TrinityCore/commit/df1ec76d54db3a91759fc93f4d6b6828d4ec0f59)

+++ Source/Game/DataStorage/Structs/F_Records.cs
@@ -96,8 +96,8 @@ namespace Game.DataStorage
     {
         public uint Id;
         public LocalizedString Reaction;
-        public uint FriendshipRepID;
-        public ushort ReactionThreshold;
+        public int FriendshipRepID;
+        public int ReactionThreshold;
         public int OverrideColor;
     }
